move struct to root.go
