+++ cmd/list.go
@@ -6,7 +6,6 @@ import (
 	"context"
 	"encoding/json"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/tools/clientcmd"
 	"k8s.io/client-go/util/homedir"
@@ -15,45 +14,7 @@ import (
 )
 
 var namespace string
-var config string
-var gvr = schema.GroupVersionResource{Group: "openrhino.org", Version: "v1alpha1", Resource: "rhinojobs"}
-
-type RhinoJobSpec struct {
-	Image string `json:"image"`
-	Parallelism *int32 `json:"parallelism,omitempty"`
-	TTL        *int32   `json:"ttl,omitempty"`
-	AppExec    string   `json:"appExec"`
-	AppArgs    []string `json:"appArgs,omitempty"`
-	DataServer string   `json:"dataServer,omitempty"`
-	DataPath   string   `json:"dataPath,omitempty"`
-}
-
-type RhinoJobStatus struct {
-	JobStatus JobStatus `json:"jobStatus"`
-}
-type JobStatus string
-
-const (
-	Pending   JobStatus = "Pending"
-	Running   JobStatus = "Running"
-	Failed    JobStatus = "Failed"
-	Completed JobStatus = "Completed"
-)
-
-type RhinoJob struct {
-	metav1.TypeMeta   `json:",inline"`
-	metav1.ObjectMeta `json:"metadata,omitempty"`
-
-	Spec   RhinoJobSpec   `json:"spec,omitempty"`
-	Status RhinoJobStatus `json:"status,omitempty"`
-}
-
-type RhinoJobList struct {
-	metav1.TypeMeta `json:",inline"`
-	metav1.ListMeta `json:"metadata,omitempty"`
-	Items           []RhinoJob `json:"items"`
-}
-
+var kubeconfig string
 
 var listCmd = &cobra.Command{
 	Use:   "list",
@@ -62,19 +23,18 @@ var listCmd = &cobra.Command{
 	Example: `  rhino list
   rhino list --namespace user_func`,
 	RunE: func(cmd *cobra.Command, args []string) error{
-		var kubeconfig string
-
-		if home := homedir.HomeDir(); home != "" {
-			kubeconfig = filepath.Join(home, ".kube", "config")
-		} else {
-			if len(config) == 0 {
+		var configPath string
+		if len(kubeconfig) == 0 {
+			if home := homedir.HomeDir(); home != "" {
+				configPath = filepath.Join(home, ".kube", "config")
+			} else {
 				fmt.Println("Error: kubeconfig file not found, please use --config to specify the absolute path")
 				os.Exit(0)
 			}
-			kubeconfig = config
-		}
-
-		config, err := clientcmd.BuildConfigFromFlags("", kubeconfig)
+		} else {
+			configPath = kubeconfig
+		}		
+		config, err := clientcmd.BuildConfigFromFlags("", configPath)
 		if err != nil {
 			return err
 		}
@@ -102,12 +62,12 @@ var listCmd = &cobra.Command{
 
 func init() {
 	rootCmd.AddCommand(listCmd)
-	listCmd.Flags().StringVarP(&namespace, "namespace", "n", "default", "namespace of the rhinojob")
-	listCmd.Flags().StringVar(&config, "config", "", "kubernetes config path")
+	listCmd.Flags().StringVarP(&namespace, "namespace", "n", "", "namespace of the rhinojob")
+	listCmd.Flags().StringVar(&kubeconfig, "kubeconfig", "", "kubernetes config path")
 }
 
 func listRhinoJob(client dynamic.Interface, namespace string) (*RhinoJobList, error) {
-	list, err := client.Resource(gvr).Namespace(namespace).List(context.TODO(), metav1.ListOptions{})
+	list, err := client.Resource(RhinoJobGVR).Namespace(namespace).List(context.TODO(), metav1.ListOptions{})
 	if err != nil {
 			return nil, err
 	}
